Changes after Greg review
