+++ kong2kic/kong2kic_test.go
@@ -45,56 +45,56 @@ func Test_convertKongGatewayToIngress(t *testing.T) {
 		{
 			name:           "Kong to KIC: kic v3.x Gateway API, yaml",
 			inputFilename:  "input.yaml",
-			outputFilename: "kicv3_gateway/yaml/output-expected.yaml",
+			outputFilename: "kicv3_gateway/output-expected.yaml",
 			builderType:    KICV3GATEWAY,
 			wantErr:        false,
 		},
 		{
 			name:           "Kong to KIC: kic v3.x Gateway API, json",
 			inputFilename:  "input.yaml",
-			outputFilename: "kicv3_gateway/json/output-expected.json",
+			outputFilename: "kicv3_gateway/output-expected.json",
 			builderType:    KICV3GATEWAY,
 			wantErr:        false,
 		},
 		{
 			name:           "Kong to KIC: kic v3.x Ingress API, yaml",
 			inputFilename:  "input.yaml",
-			outputFilename: "kicv3_ingress/yaml/output-expected.yaml",
+			outputFilename: "kicv3_ingress/output-expected.yaml",
 			builderType:    KICV3INGRESS,
 			wantErr:        false,
 		},
 		{
 			name:           "Kong to KIC: kic v3.x Ingress API, json",
 			inputFilename:  "input.yaml",
-			outputFilename: "kicv3_ingress/json/output-expected.json",
+			outputFilename: "kicv3_ingress/output-expected.json",
 			builderType:    KICV3INGRESS,
 			wantErr:        false,
 		},
 		{
 			name:           "Kong to KIC: kic v2.x Gateway API, yaml",
 			inputFilename:  "input.yaml",
-			outputFilename: "kicv2_gateway/yaml/output-expected.yaml",
+			outputFilename: "kicv2_gateway/output-expected.yaml",
 			builderType:    KICV2GATEWAY,
 			wantErr:        false,
 		},
 		{
 			name:           "Kong to KIC: kic v2.x Gateway API, json",
 			inputFilename:  "input.yaml",
-			outputFilename: "kicv2_gateway/json/output-expected.json",
+			outputFilename: "kicv2_gateway/output-expected.json",
 			builderType:    KICV2GATEWAY,
 			wantErr:        false,
 		},
 		{
 			name:           "Kong to KIC: kic v2.x Ingress API, yaml",
 			inputFilename:  "input.yaml",
-			outputFilename: "kicv2_ingress/yaml/output-expected.yaml",
+			outputFilename: "kicv2_ingress/output-expected.yaml",
 			builderType:    KICV2INGRESS,
 			wantErr:        false,
 		},
 		{
 			name:           "Kong to KIC: kic v2.x Ingress API, json",
 			inputFilename:  "input.yaml",
-			outputFilename: "kicv2_ingress/json/output-expected.json",
+			outputFilename: "kicv2_ingress/output-expected.json",
 			builderType:    KICV2INGRESS,
 			wantErr:        false,
 		},
